size_t seems to break Mac compilation - anyway uintt6 is enought for messages

+++ trunk/src/ECSocket.cpp
@@ -27,7 +27,7 @@ wxString ECSocket::SendRecvMsg(const wxChar *msg) {
 
 	wxString response("");
 
-  	size_t len  = (wxStrlen(msg) + 1) * sizeof(wxChar);
+  	uint16 len  = (wxStrlen(msg) + 1) * sizeof(wxChar);
 
 	this->SetFlags(wxSOCKET_WAITALL);
 	this->Write(len);
